Support a "charge limit" number
When setting target SOC, the underlying API allows the charge limit to
also be specified. Create a number which allows this to be managed.

The current implementation is a bit messy, and is not ready to merge.
While this works correctly when modifying the charge limit (that is,
it respects the current target SOC), the reverse it not true; when
the target SOC is modified, the charge limit is lost.